feat: add nginx rule to fetch carbonio-files-public-folder-ui bundle
Added two rules to fetch the carbonio-files-public-folder-ui bundle
necessary to access a carbonio-files public folder without
authentication.

The first rule is tied to the /files/public/link/access/assets/ endpoint and it
allows to download all the assets required by the bundle.

The second rule is tied to the /files/public/link/access/ endpoint and
it downloads the index.html of the bundle.

Ref: FILES-734